Build fix and warning fixes.

+++ src/osgEarthDrivers/model_feature_stencil/FeatureStencilModelSource.cpp
@@ -263,7 +263,7 @@ createVolume(osgEarth::Symbology::Geometry* geom,
             (*top_verts)[top_vert_ptr++] = extrude_vec;
             (*bottom_verts)[bottom_vert_ptr++] = *m;
              
-            part_len += wall_vert_ptr > wall_part_ptr?
+            part_len += wall_vert_ptr > (int)wall_part_ptr?
                 (extrude_vec - (*verts)[wall_vert_ptr-2]).length() :
                 0.0;
 
@@ -279,7 +279,7 @@ createVolume(osgEarth::Symbology::Geometry* geom,
         // close the wall if it's a ring/poly:
         if ( part->getType() == osgEarth::Symbology::Geometry::TYPE_RING || part->getType() == osgEarth::Symbology::Geometry::TYPE_POLYGON )
         {
-            part_len += wall_vert_ptr > wall_part_ptr?
+            part_len += wall_vert_ptr > (int)wall_part_ptr?
                 ((*verts)[wall_part_ptr] - (*verts)[wall_vert_ptr-2]).length() :
                 0.0;
